Refactor: Simplify Stuff
